+++ app/src/main/java/com/cloudsource/quiz_android/GetJson.java
@@ -36,69 +36,5 @@ public class GetJson extends AppCompatActivity {
         request.execute();
     }
 
-    private class requestToJson extends AsyncTask<String, Void, Boolean> {
-        CategoryModel categorymodel;
-        String urlstr;
 
-        public String getUrl() {
-            return url;
-        }
-
-        public void setUrl(String url) {
-            this.urlstr = url;
-        }
-
-        @Override
-        protected void onPreExecute() {
-            super.onPreExecute();
-
-        }
-
-        @Override
-        protected Boolean doInBackground(String... strings) {
-            JSONObject object = JsonParser.getData(urlstr);
-
-            Log.i("mylog", "object" + object);
-
-            List<Category> modelList = new ArrayList<Category>();
-
-            if (object != null) {
-                try {
-                    JSONArray jsonArray = object.getJSONArray("category");
-                    for (int i = 0; i < jsonArray.length(); i++) {
-                        JSONObject innerObject = jsonArray.getJSONObject(i);
-                        Category category;
-                        category = new Category();
-                        category.setName(innerObject.getString("name"));
-                        modelList.add(category);
-                    }
-
-                    categorymodel = new CategoryModel();
-                    categorymodel.setCategories(modelList);
-
-                    String result = categorymodel.getCategories().get(1).getName();
-                    Log.i("mylog", result);
-
-
-                } catch (JSONException e) {
-                    e.printStackTrace();
-                }
-            }
-
-            return null;
-        }
-
-        @Override
-        protected void onPostExecute(Boolean aBoolean) {
-            ListviewAdapter adapter = new ListviewAdapter(getApplicationContext(), categorymodel.getCategories());
-            lv.setAdapter(adapter);
-
-            lv.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-                @Override
-                public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
-                    startActivity(new Intent(getApplicationContext(), QuizShow.class));
-                }
-            });
-        }
-    }
 }
