+++ app/src/main/java/com/cloudsource/quiz_android/QuizShow.java
@@ -15,7 +15,6 @@ import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.util.ArrayList;
-import java.util.List;
 
 /**
  * Created by ayemyathu on 8/7/16.
