+++ app/src/test/java/com/cloudsource/quiz_android/GetJsonTest.java
@@ -1,0 +1,28 @@
+package com.cloudsource.quiz_android;
+
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+/**
+ * Created by ayemyathu on 8/10/16.
+ */
+public class GetJsonTest {
+    @Before
+    public void setUp() throws Exception {
+
+    }
+
+    @After
+    public void tearDown() throws Exception {
+
+    }
+
+    @Test
+    public void onCreate() throws Exception {
+
+    }
+
+}
